move Graph and LeastCostPathTree from sbb-contribs to core
- rename Graph to SpeedyGraph
- move into package org.matsim.core.router.speedy
